put readme under module
