+++ module/README.md
@@ -1,4 +1,10 @@
-# readme ([jsr.io](https://jsr.io/@tugrulates/readme))
+# module ([jsr.io](https://jsr.io/@tugrulates/readme))
+
+Utilities for managing the modules in this repository.
+
+## Functions
+
+### [generateReadme](https://jsr.io/@tugrulates/readme/doc/~/generateReadme)
 
 Generate a README.md file for a module.
 
@@ -7,7 +13,3 @@ for each module.
 
 Alternatively, the `generateReadme` function can be used to generate a README.md
 file for a module.
-
-## Functions
-
-### [generateReadme](https://jsr.io/@tugrulates/readme/doc/~/generateReadme)
