+++ module/mod.ts
@@ -1,0 +1,7 @@
+/**
+ * Utilities for managing the modules in this repository.
+ *
+ * @module
+ */
+
+export * from "./readme.ts";
